Extract Form content into separate views so sheet states live above the Form
FeedEditSheet and AddFeedView are now thin wrappers that own only
the sheet-related @State vars and .sheet modifiers. The Form content,
form-specific states, and methods move into FeedEditForm and
AddFeedForm respectively. This prevents sheet state changes from
re-evaluating the Form body.

AddFeedForm handles post-dismiss logic for login sheets via
onChange(of:) on the boolean bindings.
